made some fixes in PR
