+++ src/components/Profile.css
@@ -1,97 +1,94 @@
-#addBtn{
-    background-color: rgb(128, 128, 128);
-    color:rgb(255, 255, 255);
-    height: 100px;
-    width: 100px;
-    font-size: 5rem;
-    border: none;
-    border-radius: 100%;
-    cursor: pointer;
-    transition: 0.5s;
-    margin: 20px;
-  
+.posts {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 25px;
+  justify-content: flex-start;
+  align-items: flex-start;
+  padding: 25px;
 }
 
-#addBtn:hover{
-        background-color: rgb(164, 164, 164);
-        transform: scale(1.2);
-    }
+.postWrapper {
+  position: relative;
+  width: 300px;
+  border-radius: 20px;
+  overflow: hidden;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+  background: white;
+  text-align: center;
+}
 
-    .postWrapper {
-        display: inline-block;
-        margin: 20px;   
-    }
-    
-    .postWrapper img {
-        display: block;
-        width: 100%; /* Ensures the image takes the full width of the container */
-        border-radius: 12px 12px 0 0;
-        max-width: 300px; /* Adjust this based on your image size */
-    }
-    
-    .misc {
-        width: 100%; /* Makes sure it matches the image width */
-        max-width: 300px; /* Ensure it doesn't exceed the image's max-width */
-        background-image: linear-gradient(to bottom, rgb(172, 172, 172), rgb(223, 223, 223));
-        border-radius: 0 0 12px 12px;
-        margin-top:-23px
-    }
-    
-    .deleteBtn{
-        float: right;
-        position: relative;
-        border: none;
-        color: white;
-        background-color: red;
-        border-radius: 100%;
-        width: 30px;
-        height: 30px;
-        cursor: pointer;
-        font-size: large;
-        right: 20px;
-        bottom: 10px;
-    }
+.postWrapper img {
+  width: 100%;
+  height: 180px;
+  object-fit: cover;
+  border-top-left-radius: 20px;
+  border-top-right-radius: 20px;
+}
 
-    .btnWrapper{
-        display: flex;
-        justify-content: space-between;
-        width: 50%;
-        margin-right: 25%;
-    }
+.misc {
+  padding: 15px;
+}
 
-    .btnWrapper button{
-        border: none;
-        font-size: large;
-        width: 30%;
-        border-radius: 45px;
-        transition: 0.5s;
-        box-shadow: rgba(0, 0, 0, 0.24) 0px 3px 8px;
-    }
+h4 {
+  font-size: 18px;
+  font-weight: bold;
+  margin: 8px 0;
+}
 
-    .btnWrapper button:hover{
-        transform: scale(1.2);
-    }
+.descriptionSpan {
+  font-size: 16px;
+  color: #666;
+}
 
-    .editBtn{
-        float: right;
-        position: relative;
-        font-size: larger;
-        border: none;
-        background-color: transparent;
-        bottom: 10px;
-        cursor: pointer;
-    }
+.editBtn {
+  background: none;
+  border: none;
+  cursor: pointer;
+  font-size: 20px;
+  padding-bottom: 12px;
+}
 
-    .inputWrapper{
-        display: block;
-        justify-content: center;
-    }
+.deleteBtn {
+  position: absolute;
+  top: 8px;
+  right: 8px;
+  background: red;
+  color: white;
+  border: none;
+  width: 30px;
+  height: 30px;
+  border-radius: 50%;
+  cursor: pointer;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-size: 18px;
+}
 
-    .inputWrapper input{
-        border: none;
-        background-color: rgb(230, 230, 230);
-        box-shadow: rgba(0, 0, 0, 0.24) 0px 3px 8px;
-        border-radius: 45px;
-        font-size: medium;
-        padding: 5px;
-    }
+#addBtn {
+  width: 120px;
+  height: 120px;
+  border-radius: 50%;
+  background: #ddd;
+  font-size: 56px;
+  border: none;
+  cursor: pointer;
+}
+
+.inputWrapper {
+  display: block;
+  justify-content: center;
+}
+
+.inputWrapper input {
+  border: none;
+  background-color: rgb(230, 230, 230);
+  box-shadow: rgba(0, 0, 0, 0.24) 0px 3px 8px;
+  border-radius: 45px;
+  font-size: medium;
+  padding: 5px;
+}
+
+.divToClick{
+    cursor: pointer;
+}
