+++ src/components/RatingStars.css
@@ -1,0 +1,31 @@
+.star-rating {
+    display: flex;
+    align-items: center;
+  }
+  
+  .star {
+    cursor: pointer;
+    margin-right: 8px;
+    transition: transform 0.3s ease-in-out, color 0.3s ease-in-out;
+  }
+  
+  .star-icon {
+    font-size: 30px;
+    color: #ccc;
+    transition: color 0.3s ease-in-out;
+  }
+  
+  .star:hover .star-icon {
+    transform: scale(1.2);
+    transition: ease-in-out 0.3s;
+    color: #ffcc00;
+  }
+  
+  .star-rating .star:hover .star-icon {
+    color: #ffcc00;
+  }
+  
+  .star-rating .star .star-icon {
+    color: #ffcc00;
+  }
+
